python-backend/requirements.txt
openai-agents
openai-chatkit
openai-guardrails
python-dotenv
pydantic
fastapi
uvicorn
